docs(initiatives) grammar fix
It wasn't until I really looked at it in #147 that I noticed how weird this sentence was structured.

The big change was needed because "we" start off as the focal point of the sentence, but then by the end of it we are being referred to as everyone (should be every one) of us individually. I can work on rewriting so that we don't lose the "big change" aspect, but I think huge impact works, because the very next sentence is the direct impact on us, what we change in our thinking.